extract Project entity to own file
