+++ src/main/scala/swiss/dasch/domain/ProjectRepository.scala
@@ -1,17 +1,11 @@
 package swiss.dasch.domain
 
-import eu.timepit.refined.api.{Refined, RefinedTypeOps}
 import io.getquill.*
 import io.getquill.jdbczio.*
 import zio.{Clock, IO, ZIO, ZLayer}
 
 import java.sql.SQLException
 import java.time.Instant
-
-type ProjectId = PositiveInt
-object ProjectId extends RefinedTypeOps[ProjectId, Int]
-
-final case class Project(id: ProjectId, shortcode: ProjectShortcode)
 
 type DbTask[A] = IO[SQLException, A]
 
@@ -23,7 +17,7 @@ final case class ProjectRepository(quill: Quill.Postgres[SnakeCase]) {
   private inline def queryProject = quote(querySchema[ProjectRow](entity = "project"))
 
   private def toProject(row: ProjectRow): Project =
-    Project(ProjectId.unsafeFrom(row.id), ProjectShortcode.unsafeFrom(row.shortcode))
+    Project(ProjectId.unsafeFrom(row.id), ProjectShortcode.unsafeFrom(row.shortcode), row.createdAt)
 
   def findByShortcode(shortcode: ProjectShortcode): DbTask[Option[Project]] =
     run(queryProject.filter(prj => prj.shortcode == lift(shortcode.value))).map(_.map(toProject)).map(_.headOption)
@@ -32,7 +26,7 @@ final case class ProjectRepository(quill: Quill.Postgres[SnakeCase]) {
     now   <- Clock.instant
     row    = ProjectRow(0, shortcode = shortcode.value, createdAt = now)
     newId <- run(queryProject.insertValue(lift(row)).returningGenerated(_.id))
-  } yield Project(ProjectId.unsafeFrom(newId), shortcode)
+  } yield Project(ProjectId.unsafeFrom(newId), shortcode, now)
 
   def deleteProject(id: ProjectId): DbTask[Long] =
     run(queryProject.filter(prj => prj.id == lift(id.value)).delete).debug
